feat: Refactor the kicker ready for templat able changes
