+++ projects/Mallard/.gitignore
@@ -62,3 +62,6 @@ buck-out/
 
 # Version info
 src/version-info.json
+
+#jest files
+coverage
