+++ projects/Mallard/package.json
@@ -7,7 +7,8 @@
         "convert-svg": "node ./scripts/svgs.js",
         "postinstall": "yarn pre-build",
         "start": "yarn pre-build && node node_modules/react-native/local-cli/cli.js start",
-        "test": "jest",
+        "test": "jest --coverage",
+        "test:watch": "jest --watch",
         "shake-android": "echo '👋 📱' && adb shell input keyevent 82 && echo '🙌📲🙌'",
         "wifi-android": "yarn pre-build && ip=$(adb shell ifconfig wlan0  | awk '/inet addr/{print substr($2,6)}') && adb tcpip 5555 && adb connect $ip:5555",
         "run-android": "yarn pre-build && adb reverse tcp:8081 tcp:8081 && adb reverse tcp:8097 tcp:8097 && adb reverse tcp:3131 tcp:3131 && react-native run-android",
