+++ projects/Mallard/src/components/article/article-header/article-header.tsx
@@ -3,44 +3,17 @@ import { View, StyleSheet, Animated } from 'react-native'
 import { HeadlineText, HeadlineKickerText } from 'src/components/styled-text'
 import { metrics } from 'src/theme/spacing'
 import { useArticleAppearance } from 'src/theme/appearance'
-import { ArticleImage } from './article-image'
+import { ArticleImage } from '../article-image'
 import { getNavigationPosition } from 'src/helpers/positions'
-import { Image } from '../../common'
-
-interface Style {
-    /* kicker */
-    kicker: {}
-    /* outer container around the header. For spacing and background colour*/
-    background: {}
-    /* text styles for the headline `<Text>` element. Mainly for colours*/
-    headline: {}
-    /* optional container around the headline `<Text>` that adds a background colour*/
-    textBackground: {}
-}
+import { newsHeaderStyles, longReadHeaderStyles } from './article-header-styles';
+import { ArticleKicker } from './article-kicker';
+import { Image } from '../../../common'
 
 export interface PropTypes {
     headline: string
     kicker?: string | null
     image?: Image | null
 }
-
-const newsHeaderStyles: StyleSheet.NamedStyles<Style> = StyleSheet.create({
-    background: {
-        alignItems: 'flex-start',
-        paddingHorizontal: metrics.horizontal,
-        paddingBottom: metrics.vertical,
-    },
-    kicker: {
-        paddingBottom: metrics.vertical / 2,
-        marginBottom: metrics.vertical / 4,
-        borderBottomWidth: StyleSheet.hairlineWidth,
-        width: '100%',
-    },
-    textBackground: {},
-    headline: {
-        marginRight: metrics.horizontal * 2,
-    },
-})
 
 const NewsHeader = ({ headline, image, kicker }: PropTypes) => {
     const { appearance } = useArticleAppearance()
@@ -56,13 +29,7 @@ const NewsHeader = ({ headline, image, kicker }: PropTypes) => {
                     image={image}
                 />
             ) : null}
-            {kicker ? (
-                <View style={[newsHeaderStyles.kicker, appearance.backgrounds]}>
-                    <HeadlineKickerText style={[appearance.text]}>
-                        {kicker}
-                    </HeadlineKickerText>
-                </View>
-            ) : null}
+            {kicker ? <ArticleKicker kicker={kicker} /> : null}
             <Animated.View
                 style={[
                     navigationPosition && {
@@ -87,25 +54,6 @@ const NewsHeader = ({ headline, image, kicker }: PropTypes) => {
     )
 }
 
-const longReadHeaderStyles: StyleSheet.NamedStyles<Style> = StyleSheet.create({
-    background: {
-        flexShrink: 0,
-        alignItems: 'flex-start',
-        justifyContent: 'flex-end',
-        height: 500,
-        marginTop: -10,
-    },
-    kicker: {
-        ...newsHeaderStyles.kicker,
-    },
-    headline: {},
-    textBackground: {
-        padding: metrics.horizontal,
-        paddingVertical: metrics.vertical / 2,
-        marginEnd: metrics.horizontal * 2,
-    },
-})
-
 const LongReadHeader = ({ headline, image, kicker }: PropTypes) => {
     const { appearance } = useArticleAppearance()
     return (
@@ -122,18 +70,7 @@ const LongReadHeader = ({ headline, image, kicker }: PropTypes) => {
                     appearance.backgrounds,
                 ]}
             >
-                {kicker ? (
-                    <View
-                        style={[
-                            longReadHeaderStyles.kicker,
-                            appearance.backgrounds,
-                        ]}
-                    >
-                        <HeadlineKickerText style={[appearance.text]}>
-                            {kicker}
-                        </HeadlineKickerText>
-                    </View>
-                ) : null}
+                {kicker ? <ArticleKicker kicker={kicker} type='longRead' /> : null}
                 <HeadlineText
                     style={[
                         longReadHeaderStyles.headline,
